+++ lib/src/contacts/model/groups.dart
@@ -4,7 +4,7 @@ class Groups {
   int totalCount;
 
   /// URL which can be used to retrieve list of groups contact belongs to.
-  int href;
+  String href;
 
   /// Constructor.
   Groups({this.totalCount, this.href});
@@ -12,7 +12,9 @@ class Groups {
   /// Construct a [Groups] object from a [json] object.
   factory Groups.fromJson(Map<String, dynamic> json) => (json == null)
       ? null
-      : Groups(totalCount: json['totalCount'], href: json['href']);
+      : Groups(
+          totalCount: int.parse(json['totalCount']),
+          href: json['href'].toString());
 
   /// Get a json object representing the [Groups]
   Map<String, dynamic> toJson() => {'totalCount': totalCount, 'href': href};