+++ lib/src/conversations/api_conversations_service.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:http/http.dart' show Response;
 import 'package:messagebird_dart/src/conversations/model/conversation.dart';
+import 'package:messagebird_dart/src/messages/model/message.dart';
 
 import '../base_service.dart';
 import 'conversations_service.dart';
@@ -35,24 +36,27 @@ class ApiConversationsService extends BaseService
   }
 
   @override
-  Future<Conversation> read(String id) {
-    get(
-      '/v1/conversations/$id',
-      hostname: BaseService.conversationsEndpoint,
-    );
+  Future<Conversation> read(String id) async {
+    final response = await get('/v1/conversations/$id',
+        hostname: BaseService.conversationsEndpoint);
+    return Future.value(
+        Conversation.fromJson(json.decode(response.body)['data']));
   }
 
   @override
   Future<ConversationMessage> readMessage(String id) async {
     final response = await get('/v1/messages/$id',
         hostname: BaseService.conversationsEndpoint);
-    return Future.value();
+    return Future.value(
+        ConversationMessage.fromJson(json.decode(response.body)['data']));
   }
 
   @override
-  Future<Response> reply(String id, Map<String, String> parameters) =>
-      post('/v1/conversations/$id/messages',
-          hostname: BaseService.conversationsEndpoint, body: parameters);
+  Future<Message> reply(String id, Message message) async {
+    final response = await post('/v1/conversations/$id/messages',
+        hostname: BaseService.conversationsEndpoint, body: message.toJson());
+    return Future.value(Message.fromJson(json.decode(response.body)['data']));
+  }
 
   @override
   Future<MessageResponse> send(ConversationMessage message) async {
@@ -63,18 +67,21 @@ class ApiConversationsService extends BaseService
   }
 
   @override
-  Future<Response> start(Map<String, dynamic> parameters) =>
-      post('/v1/conversations/start',
-          hostname: BaseService.conversationsEndpoint, body: parameters);
+  Future<Conversation> start(ConversationMessage message) async {
+    final response = await post('/v1/conversations/start',
+        hostname: BaseService.conversationsEndpoint, body: message.toJson());
+    return Future.value(
+        Conversation.fromJson(json.decode(response.body)['data']));
+  }
 
   @override
-  Future<Conversation> update(String id, ConversationStatus status) {
-    final Map<String, dynamic> json = {
+  Future<Conversation> update(String id, ConversationStatus status) async {
+    final Map<String, dynamic> j = {
       'status': status.toString().replaceAll('ConversationStatus.', '')
     };
-    final response = patch('/v1/conversations/$id',
-        hostname: BaseService.conversationsEndpoint, body: json);
+    final response = await patch('/v1/conversations/$id',
+        hostname: BaseService.conversationsEndpoint, body: j);
     return Future.value(
-        /*Conversation.fromJson(json.decode(response.body)[data])*/);
+        Conversation.fromJson(json.decode(response.body)['data']));
   }
 }