+++ lib/src/groups/groups_service.dart
@@ -1,26 +1,30 @@
-import 'package:http/http.dart' show Response;
+import 'model/group.dart';
 
 /// Groups service interface.
 abstract class GroupsService {
-  /// Adds anywhere from 1 to 50 contacts to a group.
-  Future<Response> addContacts(String groupId, List<String> contactIds);
+  /// Add contact to group. The IDs of the group and contact need to be
+  /// supplied.
+  Future<void> addContacts(String groupId, List<String> contactIds);
 
-  /// Creates a new Group. [parameters] is optional.
-  Future<Response> create(String name, {Map<String, dynamic> parameters});
+  /// Remove contact to group. The IDs of the group and contact need to be
+  /// supplied.
+  Future<void> removeContacts(String groupId, List<String> contactIds);
 
-  /// Lists existing groups. Pagination is optional. If a [limit] is set, an
+  /// Create a new Group. Parameters are optional.
+  Future<Group> create(Group group);
+
+  /// List existing groups. Pagination is optional. If a [limit] is set, an
   /// [offset] is also required.
-  Future<Response> list({int limit, int offset});
+  Future<List<Group>> list({int limit, int offset});
 
   /// View an existing group
-  Future<Response> read(String id);
+  Future<Group> read(String id);
 
-  /// Removes an existing group. The Future completes with an error if
+  /// Remove an existing group. The Future completes with an error if
   /// applicable, but the data will never contain anything meaningful as the
   /// API returns an empty response for successful deletes.
-  Future<Response> remove(String id);
+  Future<void> remove(String id);
 
-  /// Updates an existing contact. [parameters] is optional.
-  Future<Response> update(String id, String name,
-      {Map<String, dynamic> parameters});
+  /// Update an existing contact. Parameters are optional.
+  Future<Group> update(String id, {Group group});
 }